Move `integrityLevel` function to `IntegrityLevel`
