+++ appOPHD/MapObjects/Structure.cpp
@@ -23,24 +23,6 @@
 
 namespace
 {
-	struct IntegrityTransition
-	{
-		int value;
-		IntegrityLevel integrityLevel;
-	};
-
-
-	constexpr std::array<IntegrityTransition, 6> IntegrityTransitions =
-	{{
-		{100, IntegrityLevel::Mint},
-		{80, IntegrityLevel::Good},
-		{35, IntegrityLevel::Worn},
-		{20, IntegrityLevel::Decayed},
-		{1, IntegrityLevel::Unstable},
-		{0, IntegrityLevel::Destroyed},
-	}};
-
-
 	const std::map<StructureState, std::string> StructureStateDescriptions =
 	{
 		{StructureState::UnderConstruction, "Under Construction"},
@@ -49,16 +31,6 @@ namespace
 		{StructureState::Disabled, "Disabled"},
 		{StructureState::Destroyed, "Destroyed"},
 	};
-
-
-	IntegrityLevel integrityLevel(int integrity)
-	{
-		for (const auto& integrityTransition : IntegrityTransitions)
-		{
-			if (integrity >= integrityTransition.value) { return integrityTransition.integrityLevel; }
-		}
-		throw std::runtime_error("Bad integrity value: " + std::to_string(integrity));
-	}
 }
